requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic<2.0
python-dotenv